+++ forui/lib/src/widgets/select/multi/select.dart
@@ -804,6 +804,8 @@ abstract class _FMultiSelectState<S extends FMultiSelect<T>, T> extends State<S>
                 useViewPadding: widget.contentUseViewPadding,
                 useViewInsets: widget.contentUseViewInsets,
                 offset: widget.contentOffset,
+                // Tab should move to the next control instead of into the content.
+                traversalGrouped: false,
                 hideRegion: widget.contentHideRegion,
                 groupId: widget.contentGroupId,
                 cutout: widget.contentCutout,